Update encode decode unit tests
